requirements.txt
pandas
numpy
networkx
matplotlib
python-bidi
arabic_reshaper
